requirements.txt
flask
flask-cors
easyocr
edge-tts
pytesseract
Pillow
requests
base64
